Worked on rating, homepage redirect from login and rating(decimal)

+++ app/Models/Rating.php
@@ -24,7 +24,7 @@ class Rating extends Model
     ];
 
     protected $casts = [
-        'rating' => 'integer'
+        'rating' => 'float'
     ];
 
     public function job()
